Evil Twin Detector
An Evil Twin Attack is one of well-known attack since two decades, the attacker by impersonating as Legitimate AP cheats user to get connection with fake access point while he/she is not aware of that and he/she is getting connected to the hacker’s AP.
As soon as client established a connection with this fraud AP his/her all private information is lost or modified. Like when a client wants to establish a connection with a special network that is visible in his/her network preferred list of wireless network connection window, the user sends a probe message to its demanded network in wireless network connection list. This probe message is interrupt by an adversary and set up an ET attack 
So our program is an individual solution of these types of attacks, our implemented program is consist 3 main files to recognize, alert, and prevent the Evil Twin Access Point.